chore: move TimelockConvertor interface to sdk pkg
This ensures all interfaces that sdk implementations must implement are
in the sdk pkg.
