+++ .mockery.yaml
@@ -14,14 +14,5 @@ mockname: "{{.InterfaceName}}"
 inpackage: false
 outpkg: mocks
 packages:
-  github.com/smartcontractkit/mcms/internal/core/proposal/timelock:
-    config:
-      # Override the directory to place mocks inside `mocks` directory of the interface package.
-      dir: "{{.InterfaceDir}}/mocks"
-      # Ensure each mock struct name remains as the interface name.
-      mockname: "{{.InterfaceName}}"
-    interfaces:
-      TimelockConverter:
-
   github.com/smartcontractkit/mcms/sdk:
   github.com/smartcontractkit/mcms/sdk/evm:
